+++ src/protocols/handshake.rs
@@ -19,9 +19,8 @@ where
 {
     /// Prepares the node to perform specified network handshakes.
     async fn enable_handshake(&self) {
-        let (from_node_sender, mut from_node_receiver) = mpsc::channel::<ReturnableConnection>(
-            self.node().config().protocol_handler_queue_depth,
-        );
+        let (from_node_sender, mut from_node_receiver) =
+            mpsc::unbounded_channel::<ReturnableConnection>();
 
         // Use a channel to know when the handshake task is ready.
         let (tx, rx) = oneshot::channel::<()>();
@@ -83,11 +82,11 @@ where
 }
 
 /// The handler object dedicated to the [`Handshake`] protocol.
-pub struct HandshakeHandler(mpsc::Sender<ReturnableConnection>);
+pub struct HandshakeHandler(mpsc::UnboundedSender<ReturnableConnection>);
 
 impl HandshakeHandler {
-    pub(crate) async fn trigger(&self, item: ReturnableConnection) {
-        if self.0.send(item).await.is_err() {
+    pub(crate) fn trigger(&self, item: ReturnableConnection) {
+        if self.0.send(item).is_err() {
             unreachable!(); // protocol's task is down! can't recover
         }
     }